final aulas - stream

+++ banco/src/com/algaworks/banco/modelo/pagamento/DocumentoPagavel.java
@@ -1,9 +1,9 @@
 package com.algaworks.banco.modelo.pagamento;
 
-import java.util.Date;
+import java.math.BigDecimal;
 
 public interface DocumentoPagavel {
-    double getValorTotal();
+    BigDecimal getValorTotal();
     boolean estaPago();
     void quitarPagamento();
